Place shared stream concerns into base class
