+++ lib/txgh/handlers.rb
@@ -2,6 +2,7 @@ module Txgh
   module Handlers
     autoload :Github,            'txgh/handlers/github'
     autoload :Response,          'txgh/handlers/response'
+    autoload :StreamResponse,    'txgh/handlers/stream_response'
     autoload :TgzStreamResponse, 'txgh/handlers/tgz_stream_response'
     autoload :Transifex,         'txgh/handlers/transifex'
     autoload :Triggers,          'txgh/handlers/triggers'
